# services/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/secrets": {
      "get": {
        "operationId": "SecretsController_list",
        "parameters": [
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Secrets"]
      },
      "put": {
        "operationId": "SecretsController_put",
        "parameters": [
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "pattern": "^[a-z0-9][a-z0-9._/-]*$"
                  },
                  "value": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 8000
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500
                  }
                },
                "required": ["name", "value"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Secrets"]
      }
    },
    "/api/v1/secrets/providers": {
      "get": {
        "operationId": "SecretsController_providers",
        "parameters": [
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Secrets"]
      }
    },
    "/api/v1/secrets/{id}/rollback": {
      "post": {
        "operationId": "SecretsController_rollback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "version": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0
                  }
                },
                "required": ["version"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Secrets"]
      }
    },
    "/api/v1/secrets/{id}": {
      "delete": {
        "operationId": "SecretsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-workspace-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["Secrets"]
      }
    },
...
    "/api/v1/organizations": {
      "get": {
        "operationId": "OrganizationsController_list",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Organizations"]
      },
      "post": {
        "operationId": "OrganizationsController_create",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 100,
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000,
                    "nullable": true
                  }
                },
                "required": ["name", "slug"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Organizations"]
      }
    },
    "/api/v1/organizations/{id}": {
      "get": {
        "operationId": "OrganizationsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Organizations"]
      },
      "patch": {
        "operationId": "OrganizationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": false,
            "description": "Workspace đang thao tác. Bắt buộc với route thuộc phạm vi workspace.",
            "schema": {
              "type": "string",
              "example": "wsp_01H..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "version": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000,
                    "nullable": true
                  },
                  "status": {
                    "type": "string",
                    "enum": ["ACTIVE", "SUSPENDED", "ARCHIVED"]
                  }
                },
                "required": ["version"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Organizations"]
      }
    },
...
  },
  "info": {
    "title": "AI Social OS API",
    "description": "Mọi route đều cần đăng nhập trừ khi ghi rõ là công khai. Route thuộc phạm vi workspace cần thêm header `X-Workspace-Id`; thiếu nó là 400, và trỏ vào workspace không phải của mình là 404 chứ không phải 403 — 403 sẽ xác nhận workspace đó có tồn tại.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}